Added integration tests

+++ p05_cargo_pg/codar/src/math/simple.rs
@@ -1,4 +1,4 @@
-pub fn add(args: &[i32]) -> i32 {
+pub fn add(args: &[i64]) -> i64 {
     let mut result = 0;
     for arg in args.iter() {
         result += arg;
